VocaDB front-end only development. This is a minimal version of the Development environment guide that instructs the user how to set up local development for VocaDB for the front-end only. This bypasses the need to set up the back-end, server and database locally and uses the beta website to be able to quickly start developing. Guide should be OS agnostic.
This information is currently only available by searching through the discord history, which isn't a very straightforward process. It would serve better to be put on the wiki.

I've also taken the liberty to fix some stray typos in both the existing Development environment documents and correct the incorrect `npm run dev` command in the windows dev env document to `npm run development`.